Create ISSET.php
This is without a doubt my favorite technique in PHP, it is a REAL time saver and space saver. Instead of having ALL your code and functions on seperate pages, IE, code to send something, second page to retreive the code and do something, AKA a page to add a blog, then a page to insert a blog into the database, you basically combine those pages, but you are looking for very specific conditions to be met, before any code at all is rendered to the browser, and since its PHP it wont even be shown anyway until its called to. 
This is REALLY simple, basically it looks for input from submit buttons, in a normal submit button you have: <input name=“blah” type=“submit” value=“blah” /> the normal code looks for a Name, a type, and a value. We all know that type is what kind of button it is, and value is what gets printed to the button, but in most cases people just call their submit buttons like “submit1” or whatever, but I beg you to think of specific names, specific to the purpose of the button, so we’ll call ours for this code example “Edit”

You can also use UNSET to remove things from the page, like say you want to abort the edit you were about to make to something in the database. So you would have a button with the name of “Abort_Edit” and voila. when it finds the value “Abort_Edit” in the $_POST superglobal it knows to remove the code “Edit” really quite simple. I have also used $first_view for alot of my admin functions, so that if nothing has been sent to the browser from any functions, $first_view==“1” then I have it set to display the first view code only if first view is present (typically a list of things to view or edit, could be WHATEVER you want. but when you go to use the Add_Blog section, looking to where I have unset the edit, you would change that to just simple say unset ($first_view) the code then only renders the first view section if its set, right?..lol