Create Switch Conditions.php
Another of my favorite things to do in PHP would be switch conditions. It is a really nice way of making IF statements even less complicated, especially if you are only checking one variable, for a variety of things it could be. This specific example comes from the Ayers Farmers Market site I worked on while at BCS, we needed a way to upload images into specific folders, as you can see, the switch is testing $category, to find out if it eqauls the case, any number of things, in the first group of four its looking for it to be one of the four seasons, if it finds that, the the $destination_file is eqaul to a specific folder location (this was part of the process code for an FTP upload script, for uploading images.) when it would find a case to be correct it would break and stop executing along that line of code. Its really easy, think of it kind of like a windows start menu, the switch being “Program” the cases being what kind of program you need, so for example, I am looking for a graphic program on my computer I would have several options, fireworks, photoshop, illustrator, etc. Thats probably a bit more complicated than you needed, suffice it to say that to do the following code with IF.. ELSEIF statements actually gave me a migraine, but as soon as I used switch conditions my headache let up